address initial feedback.
